'reply'

+++ src/main/java/com/majiang/community/controller/QuestionController.java
@@ -16,7 +16,7 @@ public class QuestionController {
     QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable("id") Integer id,
+    public String question(@PathVariable("id") Long id,
                            Model model) {
         QuestionDTO questionDTO = questionService.getById(id);
         questionService.incView(id);
